requirements.txt
django
djangorestframework
pillow
python-dotenv
django-cors-headers
reportlab
python-dateutil